creating new test file for switch test and removing from network_test
new test case for switch port enable disable test and removing from
network_test case.
