+++ io/net/switch_test.py.data/README.txt
@@ -1,0 +1,20 @@
+description:
+------------------------
+This Program to test switch port enable.
+-----------------------------
+Inputs Needed To Run Tests:
+-----------------------------
+interface --> host interface to perform test.
+peer_ip --> peer interface to perform test.
+host-IP --> Specify host-IP for ip configuration.
+netmask --> specify netmask for ip configuration.
+switch_name --> switch ip address to test.
+userid --> switch userid
+password --> switch password
+port_id --> port id to perform test.
+-----------------------
+Requirements:
+-----------------------
+1. install paramiko using pip
+command: pip install paramiko
+
