+++ io/net/switch_test.py.data/switch_test.yaml
@@ -1,0 +1,8 @@
+interface:
+peer_ip:
+host_ip:
+netmask:
+switch_name:
+userid:
+password:
+port_id:
